refactor del FieldVerifier
