+++ src/controller/ClientController.java
@@ -9,7 +9,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.control.*;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
-import model.FieldVerifier;
+import util.FieldVerifier;
 import model.Gender;
 import model.Owner;
 import javax.swing.*;
