--- src/model/FieldVerifier.java
@@ -1,34 +1,0 @@
-package model;
-
-public interface FieldVerifier {
-
-    default boolean fiscalCodeVerifier(String fiscalCode){
-        return fiscalCode.matches("^([A-Za-z]{6}[0-9lmnpqrstuvLMNPQRSTUV]{2}[abcdehlmprstABCDEHLMPRST]{1" +
-                "}[0-9lmnpqrstuvLMNPQRSTUV]{2}[A-Za-z]{1}[0-9lmnpqrstuvLMNPQRSTUV]{3}[A-Za-z]{1})$|([0-9]{11})$");
-    }
-    default boolean emailVerifier(String email){
-        return email.matches("(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[" +
-                "\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-" +
-                "\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(" +
-                "?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\\.){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|" +
-                "[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\" +
-                "\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])");
-    }
-    default boolean phoneNumberVerifier(String number){
-        return number.matches("^(\\((00|\\+)39\\)|(00|\\+)39)?(38[890]|35[01]|34[7-90]|36[680]|33[3-90]|32[0-9])\\d{7}$");
-    }
-
-//    default EventHandler<KeyEvent> fieldChangeListener(final Integer max_Lengh) {
-//        return new EventHandler<KeyEvent>() {
-//            @Override
-//            public void handle(KeyEvent event) {
-//                TextField field = (TextField) event.getSource();
-//                String text = field.getText();
-//                // I need here something like:
-//                if(KeyEvent.VK_ENTER){
-//                    // do special part for ENTER KEY
-//                }
-//            }
-//        }
-//    }
-}
